requirements.txt
spotipy
openai==0.28.1
python-dotenv
